+++ mainwindow.h
@@ -27,7 +27,7 @@ private:
     Prefs _prefs;
     bool _userPressedStop = false;
     QString _previousRunFolders = QStringLiteral("");
-    ushort _previousRunThumbnails = 0;
+    int _previousRunThumbnails = 0;
 
 private slots:
     void deleteTemporaryFiles() const;
@@ -36,13 +36,13 @@ private slots:
     void loadExtensions();
     bool detectffmpeg() const;
 
-    void setComparisonMode(const short &mode) { if(mode == _prefs._PHASH) ui->selectPhash->click(); else ui->selectSSIM->click(); ui->directoryBox->setFocus(); }
-    void on_selectThumbnails_activated(const int &index) { _prefs._thumbnails = static_cast<ushort>(index); ui->directoryBox->setFocus(); }
+    void setComparisonMode(const int &mode) { if(mode == _prefs._PHASH) ui->selectPhash->click(); else ui->selectSSIM->click(); ui->directoryBox->setFocus(); }
+    void on_selectThumbnails_activated(const int &index) { _prefs._thumbnails = index; ui->directoryBox->setFocus(); }
     void on_selectPhash_clicked(const bool &checked) { if(checked) _prefs._ComparisonMode = _prefs._PHASH; ui->directoryBox->setFocus(); }
     void on_selectSSIM_clicked(const bool &checked) { if(checked) _prefs._ComparisonMode = _prefs._SSIM; ui->directoryBox->setFocus(); }
-    void on_blocksizeCombo_activated(const int &index) { _prefs._ssimBlockSize = static_cast<short>(pow(2, index+1)); ui->directoryBox->setFocus(); }
-    void on_differentDurationCombo_activated(const int &index) { _prefs._differentDurationModifier = static_cast<short>(index); ui->directoryBox->setFocus(); }
-    void on_sameDurationCombo_activated(const int &index) { _prefs._sameDurationModifier = static_cast<short>(index); ui->directoryBox->setFocus(); }
+    void on_blocksizeCombo_activated(const int &index) { _prefs._ssimBlockSize = static_cast<int>(pow(2, index+1)); ui->directoryBox->setFocus(); }
+    void on_differentDurationCombo_activated(const int &index) { _prefs._differentDurationModifier = index; ui->directoryBox->setFocus(); }
+    void on_sameDurationCombo_activated(const int &index) { _prefs._sameDurationModifier = index; ui->directoryBox->setFocus(); }
     void on_thresholdSlider_valueChanged(const int &value) { ui->thresholdSlider->setValue(value); calculateThreshold(value); ui->directoryBox->setFocus(); }
     void calculateThreshold(const int &value);
